requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
anyio==4.6.2.post1
astunparse==1.6.3
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.3.0
cycler==0.12.1
distro==1.9.0
fastapi==0.115.4
filelock==3.16.1
flatbuffers==24.3.25
fonttools==4.54.1
fsspec==2024.10.0
gast==0.6.0
google-auth==2.36.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.68.0
h11==0.14.0
h5py==3.12.1
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.26.2
idna==3.10
iniconfig==2.0.0
Jinja2==3.1.4
jiter==0.6.1
keras==3.6.0
kiwisolver==1.4.7
libclang==18.1.1
logging==0.4.9.6
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
mdurl==0.1.2
ml-dtypes==0.3.2
mpmath==1.3.0
namex==0.0.8
networkx==3.4.2
numpy==1.26.4
oauthlib==3.2.2
openai==1.53.0
opt_einsum==3.4.0
optree==0.13.1
packaging==24.1
pandas==2.2.3
pillow==11.0.0
pluggy==1.5.0
protobuf==4.25.5
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pyotp==2.9.0
pyparsing==3.2.0
pytest==8.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rsa==4.9
safetensors==0.4.5
seaborn==0.13.2
six==1.16.0
sniffio==1.3.1
starlette==0.41.2
sympy==1.13.3
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.2
tensorflow-estimator==2.14.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
tokenizers==0.20.3
torch==2.2.2
tqdm==4.66.6
transformers==4.46.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.0
Werkzeug==3.1.3
wrapt==1.14.1
